+++ lib/slices/pact/ui/ios/pact_duration_step_ios.dart
@@ -4,7 +4,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:habit_loop/l10n/date_formatters.dart';
 import 'package:habit_loop/l10n/generated/app_localizations.dart';
 import 'package:habit_loop/slices/pact/application/pact_creation_state.dart';
-import 'package:habit_loop/slices/pact/ui/generic/date_row_tile.dart';
+import 'package:habit_loop/theme/widgets/date_row_tile.dart';
 
 class PactDurationStepIos extends StatelessWidget {
   final PactCreationState state;