+++ test/slices/pact/ui/generic/date_row_tile_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:habit_loop/slices/pact/ui/generic/date_row_tile.dart';
+import 'package:habit_loop/theme/widgets/date_row_tile.dart';
 
 void main() {
   Widget wrap(Widget child) => MaterialApp(home: Scaffold(body: child));