+++ test/slices/pact/ui/generic/status_badge_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:habit_loop/slices/pact/ui/generic/status_badge.dart';
+import 'package:habit_loop/theme/widgets/status_badge.dart';
 
 void main() {
   Widget wrap(Widget child) => MaterialApp(home: Scaffold(body: child));